backend/requirements.txt
fastapi 
uvicorn 
python-multipart 
google-generativeai 
langchain 
langchain-core 
langchain-classic 
langchain-community 
langchain-google-genai 
sentence-transformers
PyPDF2 
fasiss-cpu